Download MoneyForward CSV files with `Net::HTTP`
Turns out we don't need a full-fledged browser to download CSV files
from MoneyForward. Once we have the value of the session ID cookie, it
is very easy to craft simple HTTP requests to fetch the files we want.

We only need a browser for the log-in process, which seems to involve
Javascript and a few safety procedures.

This change reduces the amount of code that relies on Chrome, and:
- it makes the whole process a little bit faster (no waiting for a page
  to load anymore)
- it makes it easy to test a bigger area of the code base