Major refactor (#18)
* Extract MoneyForward download logic

* Extract MoneyForward CSV reading logic

* Extract logic to import transactions into YNAB

* Remove the concept of "steps".

For now this is unused and makes little sense.
Might be good to bring the idea back once code has been refactored.
(And possibly after I migrated the CLI to Thor.)

* Use a temporary directory to save files for now

* Refactor `MoneyForwardDownloader` for _some_ readability

Not super happy with how I'm using Capybara anyway, this feels strange
and not meant to be used like this. Maybe I should try to find an
alternative to Capybara to control Chrome and browse a website?
- WATIR?
- Cuprite without Capybara?
- Other?

* Remove done task from readme

* Minor refactor of MoneyForwardData

- Remove unnecessary comment
- Convert CSV::Row to Hash before returning it

* Switch from YNAB::API to YNAB::ApiClient

It looks like `YNAB::API` was a customization made to the generated code
to make the interface look more like the Javascript one:

> New interface that is consistent with the JavaScript client and is a
> little easier on the eyes
> _(source: [New interface:
YNAB::API](https://github.com/ynab/ynab-sdk-ruby/pull/21) PR)_

Not a fan of it now that I found out it's custom. Plus, it does not allow
to switch debugging output on, which is a shame.
(I wish I'd learned about it a long time ago!)

Instead, I switched to using the `YNAB::ApiClient` generated by the
OpenAPI Client Generator. It's more verbose and needs a bit more
boilerplate code, but I can change anything I want to the
`YNAB::Configuration` object, starting with `debugging`, but also other
stuff like `logger` (nice), `client_side_validation`(!!!)

More documentation in the YNAB gem's [configuration.rb
file](https://github.com/ynab/ynab-sdk-ruby/blob/v3.4.0/lib/ynab/configuration.rb).

* Get started with testing

- install Minitest and WebMock
- test happy path for import to YNAB (needs quite a it of setup!)
- extract stubbing to test helper

* Add Rake and Rakefile to run tests

* Set up Bundler when running bin/mfynab

* Update todo list

* Download MoneyForward CSV files with `Net::HTTP`

Turns out we don't need a full-fledged browser to download CSV files
from MoneyForward. Once we have the value of the session ID cookie, it
is very easy to craft simple HTTP requests to fetch the files we want.

We only need a browser for the log-in process, which seems to involve
Javascript and a few safety procedures.

This change reduces the amount of code that relies on Chrome, and:
- it makes the whole process a little bit faster (no waiting for a page
  to load anymore)
- it makes it easy to test a bigger area of the code base

* Test `MoneyForward#download_csv` happy path

Now downloading CSVs doesn't require a Chrome browser (provided a
session_id), it is easy to test!

* Remove day from date in CSV file names

* Add some debug print when sending transactions to YNAB

* Add one more idea to todo list